optimizacion de la transpuesta en cuda

+++ test/laboratory.py
@@ -20,8 +20,8 @@ import sys
 
 # Modulo 2
 def generate(size):
-    A = np.random.randn(size,size).astype(np.float32)
-    xv = np.random.randn(size,1).astype(np.float32)
+    A = np.random.randn(size,size).astype(np.float64)
+    xv = np.random.randn(size,1).astype(np.float64)
     turn_dominant(A)
     B = A @ xv
     return A,B,xv
@@ -126,7 +126,7 @@ def runner(size=100, N=100, tol=0.001, alpha=0.001,
       cuda=False,opencl=False,numpy=False,
       jaco=False, grad_descent=False, conj_grad=False):
   A,B,xv = generate(size)
-  x_ini = np.ones(xv.shape).astype(np.float32)
+  x_ini = np.ones(xv.shape).astype(np.float64)
   platform = []
   method = []
   Size = []
